Lightning Talk proposal, Jürgen Riegler, SwiftUI
There were no more open dates in the course/lightning-talk calendar. 
On Nov. 14, Florian Hitzler is doing a lightning talk about Ionic. My topic would be SwiftUI (new technology for native IOS development).
So i thought it could be a good idea to hold my lightning talk on this day, for the class to have a comparison between web and native development.